refactor: Separate code
