--- src/contexts/is-revision-view.ts
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const IsRevisionViewContext = createContext<boolean>(false)
