+++ src/pages/entity/repository/compare/[entity_id]/[revision_id].tsx
@@ -2,14 +2,14 @@ import { GetStaticPaths, GetStaticProps } from 'next'
 
 import { Revision } from '@/components/author/revision/revision'
 import { FrontendClientBase } from '@/components/frontend-client-base'
-import { IsRevisionViewContext } from '@/contexts/is-revision-view'
 import { RevisionProps, RevisionPage } from '@/data-types'
 import { Instance } from '@/fetcher/graphql-types/operations'
 import { requestRevision } from '@/fetcher/revision/request'
 import { renderedPageNoHooks } from '@/helper/rendered-page'
+import { IsRevisionViewContext } from '@/serlo-editor-integration/context/is-revision-view'
 
 export default renderedPageNoHooks<RevisionProps>(({ pageData }) => (
-  // TODO: Does not do anything yet. Just to show where the context could be provided. 
+  // TODO: Does not do anything yet. Just to show where the context could be provided.
   <IsRevisionViewContext.Provider value>
     <FrontendClientBase
       entityId={pageData.revisionData.thisRevision.id}
