+++ src/serlo-editor-integration/context/is-revision-view.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react'
+
+export const IsRevisionViewContext = createContext<boolean>(false)
