+++ src/serlo-editor-integration/revision-view-extra-info.tsx
@@ -1,6 +1,9 @@
-
 // Copied from ExtraRevisionViewInfo
-export function RevisionViewExtraInfo({ children }: { children: React.ReactNode }) {
+export function RevisionViewExtraInfo({
+  children,
+}: {
+  children: React.ReactNode
+}) {
   return (
     <span className="break-all bg-editor-primary-100 px-1 text-sm">
       {children}
